# perf(es/minifier): Use efficient logic for entry tracking in DCE