--- a/crates/swc_ecma_transforms_optimization/src/simplify/dce/mod.rs
+++ b/crates/swc_ecma_transforms_optimization/src/simplify/dce/mod.rs
@@ -102,6 +102,9 @@ struct Data {
 
     /// Variable usage graph
     graph: FastDiGraphMap<usize, VarInfo>,
+    /// Entrypoints.
+    entries: Vec<usize>,
+
     graph_ix: IndexSet<Id, ahash::RandomState>,
 }